\name{AsymmLL}
\alias{AsymmLL}
\title{
...
}
\description{
...
}
\usage{
AsymmLL(m_j, c_j, a, b)
}
\arguments{
\item{m_j}{Integer for the number of alternate allele counts}
\item{c_j}{Interger for the total number of reference and alternate counts}
\item{a}{...}
\item{b}{...}
}
\value{
Invisibly the value of the likelihood
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
