\name{LL}
\alias{LL}
\title{
Likelihood of ???
}
\description{
Returns the likelihood of ???
}
\usage{
LL(m_j, c_j, n, b)
}
\arguments{
\item{m_j}{???}
\item{c_j}{???}
\item{n}{???}
\item{b}{???}
}
\note{
}
\value{
Invisibly the value of the likelihood
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
