\name{SymmLL}
\alias{SymmLL}
\title{
...
}
\description{
...
}
\usage{
SymmLL(m_j, c_j)
}
\arguments{
\item{m_j}{Integer for the number of alternate allele counts}
\item{c_j}{Interger for the total number of reference and alternate counts}
}
\value{
Invisibly the value of the likelihood
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
