\name{TotalLL}
\alias{TotalLL}
\title{
Likelihood of ???
}
\description{
Returns the likelihood of ???
}
\usage{
TotalLL(m, c, n, b, method, log)
}
\arguments{
\item{m_j}{???}
\item{c_j}{???}
\item{n}{???}
\item{b}{???}
\item{method}{???}
\item{log}{???}
}
\note{
}
\value{
Invisibly the value of the likelihood
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
