\name{SymmLL}
\alias{SymmLL}
\title{
Symmetric Log-Likelihood of mosaic mutations
}
\description{
Symmetric Log-Likelihood of mosaic mutations
}
\usage{
SymmLL(m, c, nb)
}
\arguments{
\item{m}{Integer vector of number of alternate allele counts}
\item{c}{Integer vector of total number of reference and alternate counts}
\item{nb}{Integer indicating how many branches the model includes.
	  \eqn{nb = 4} or \eqn{nb = 15}}
}
\value{
Invisibly the value of the likelihood
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
