\name{TotalLL}
\alias{TotalLL}
\title{
...
}
\description{
...
}
\usage{
TotalLL(m, c, a, b)
}
\arguments{
\item{m}{Integer vector for the number of alternate allele counts}
\item{c}{Interger vector for the total number of reference and alternate counts}
\item{a}{...}
\item{b}{...}
}
\value{
Invisibly the value of the likelihood
}
\author{
\cr
David Brown\cr
\email{brownd7@mskcc.org}\cr
}
\keyword{}
